requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2==2.9.11
python-dotenv==1.2.2
pydantic==2.12.5
pydantic[email]==2.12.5
python-jose[cryptography]==3.5.0
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.28.1
email-validator==2.2.0